+++ plone/app/upgrade/v50/testing.py
@@ -2,11 +2,23 @@ from plone.app.testing import PloneSandboxLayer
 from plone.app.testing import PLONE_FIXTURE
 from plone.testing.z2 import FunctionalTesting, login
 from zope.component.hooks import setSite
+from zope.configuration import xmlconfig
 import os
 
 
 class RealUpgradeLayer(PloneSandboxLayer):
     defaultBases = (PLONE_FIXTURE,)
+
+    def setUpZope(self, app, configurationContext):
+        # load ZCML
+        # In 5.0 alpha we install or upgrade plone.app.caching,
+        # so it must be available to Zope..
+        import plone.app.caching
+        xmlconfig.file(
+            'configure.zcml',
+            plone.app.caching,
+            context=configurationContext
+        )
 
     def setUpPloneSite(self, portal):
         app = portal.aq_parent